requirements.txt
agno==1.0.4
beautifulsoup4==4.13.0
Flask==3.1.0
markdown==3.7
linkedin-api-client==0.3.0
google-generativeai==0.8.4
newsapi==0.1.1
newsapi-python==0.2.7